chore: Move account repo test to repo directory
