+++ app/gateways/db/postgres/accounts/debit_test.go
@@ -1,4 +1,4 @@
-package postgres_test
+package accountspg
 
 import (
 	"errors"
@@ -6,20 +6,20 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/jpgsaraceni/suricate-bank/app/domain/entities/account"
-	accountspg "github.com/jpgsaraceni/suricate-bank/app/gateways/db/postgres/accounts"
 	"github.com/jpgsaraceni/suricate-bank/app/vos/cpf"
 	"github.com/jpgsaraceni/suricate-bank/app/vos/hash"
 	"github.com/jpgsaraceni/suricate-bank/app/vos/money"
 )
 
-func TestCredit(t *testing.T) {
+func TestDebit(t *testing.T) {
 	t.Parallel()
 
 	var (
-		accountId1   = account.AccountId(uuid.New())
-		accountId2   = account.AccountId(uuid.New())
-		testHash, _  = hash.NewHash("nicesecret")
-		testMoney, _ = money.NewMoney(10)
+		accountId1     = account.AccountId(uuid.New())
+		accountId2     = account.AccountId(uuid.New())
+		testHash, _    = hash.NewHash("nicesecret")
+		testMoney20, _ = money.NewMoney(20)
+		testMoney10, _ = money.NewMoney(10)
 	)
 
 	type args struct {
@@ -29,7 +29,7 @@ func TestCredit(t *testing.T) {
 
 	type testCase struct {
 		name            string
-		runBefore       func(repo *accountspg.Repository) error
+		runBefore       func(repo *Repository) error
 		args            args
 		expectedBalance int
 		err             error
@@ -37,27 +37,28 @@ func TestCredit(t *testing.T) {
 
 	testCases := []testCase{
 		{
-			name: "successfully credit 10 to account with 0 balance",
-			runBefore: func(repo *accountspg.Repository) error {
+			name: "successfully debit 10 from account with 20 balance",
+			runBefore: func(repo *Repository) error {
 				return repo.Create(
 					testContext,
 					&account.Account{
-						Id:     accountId1,
-						Name:   "Nice name",
-						Cpf:    cpf.Random(),
-						Secret: testHash,
+						Id:      accountId1,
+						Name:    "Nice name",
+						Cpf:     cpf.Random(),
+						Secret:  testHash,
+						Balance: testMoney20,
 					},
 				)
 			},
 			args: args{
 				accountId: accountId1,
-				amount:    testMoney,
+				amount:    testMoney10,
 			},
 			expectedBalance: 10,
 		},
 		{
-			name: "successfully credit 10 to account with 10 balance",
-			runBefore: func(repo *accountspg.Repository) error {
+			name: "successfully debit 20 from account with 20 balance",
+			runBefore: func(repo *Repository) error {
 				return repo.Create(
 					testContext,
 					&account.Account{
@@ -65,23 +66,23 @@ func TestCredit(t *testing.T) {
 						Name:    "Nice name",
 						Cpf:     cpf.Random(),
 						Secret:  testHash,
-						Balance: testMoney,
+						Balance: testMoney20,
 					},
 				)
 			},
 			args: args{
 				accountId: accountId2,
-				amount:    testMoney,
+				amount:    testMoney20,
 			},
-			expectedBalance: 20,
+			expectedBalance: 0,
 		},
 		{
-			name: "fail to credit inexistent account",
+			name: "fail to debit inexistent account",
 			args: args{
 				accountId: account.AccountId(uuid.New()),
-				amount:    testMoney,
+				amount:    testMoney10,
 			},
-			err: accountspg.ErrQuery,
+			err: ErrQuery,
 		},
 	}
 
@@ -90,7 +91,7 @@ func TestCredit(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 
-			repo := accountspg.NewRepository(dbPool)
+			repo := NewRepository(dbPool)
 
 			if tt.runBefore != nil {
 				err := tt.runBefore(repo)
@@ -100,12 +101,12 @@ func TestCredit(t *testing.T) {
 				}
 			}
 
-			if err := repo.CreditAccount(testContext, tt.args.accountId, tt.args.amount); !errors.Is(err, tt.err) {
-				t.Fatalf(" got error: %s expected error: %s", err, tt.err)
+			if err := repo.DebitAccount(testContext, tt.args.accountId, tt.args.amount); !errors.Is(err, tt.err) {
+				t.Fatalf("got error: %s expected error: %s", err, tt.err)
 			}
 
 			if gotBalance, _ := repo.GetBalance(testContext, tt.args.accountId); gotBalance != tt.expectedBalance {
-				t.Fatalf("got %d expected %d ", gotBalance, tt.expectedBalance)
+				t.Fatalf("got %d expected %d", gotBalance, tt.expectedBalance)
 			}
 		})
 	}
